refactor: ui관련 파일 View폴더에서 일괄 관리
